requirements.txt
amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
Django==3.1
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
environs==8.0.0
Faker==0.9.1
flake8==3.8.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
mailjet-rest==1.3.3
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.7.1
mccabe==0.6.1
mixer==6.1.3
packaging==20.4
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
